requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
matplotlib==3.8.2
qrcode==7.4.2
Pillow==10.1.0